Add unit tests for /api/users
Add more unit tests for /api/users routes. This includes failure
conditions, such as registering with the same email as a previous user
and logging in with the incorrect password.

This should be used in conjunction with a local database instance for
testing purposes. It will delete all stored data upon running `cargo
test`, before adding some test data. This can be done on the Atlas
instance, but may lead to some issues if people run tests at the same
time.

In the future, it may be easier to mock the MongoDB instance to mitigate
the need for this, however for now it should be fine to setup the
database and run some API queries on the data.