Add unit tests for /api/users (#42)
* Add some basic tests

Add some basic tests under the `tests/` directory. There are only two
tests, which setup the server and ping two endpoints, one of which
requires a connection to the MongoDB server itself.

* Add unit tests for /api/users

Add more unit tests for /api/users routes. This includes failure
conditions, such as registering with the same email as a previous user
and logging in with the incorrect password.

This should be used in conjunction with a local database instance for
testing purposes. It will delete all stored data upon running `cargo
test`, before adding some test data. This can be done on the Atlas
instance, but may lead to some issues if people run tests at the same
time.

In the future, it may be easier to mock the MongoDB instance to mitigate
the need for this, however for now it should be fine to setup the
database and run some API queries on the data.

* Move dotenv::dotenv() back to main

Move the environment populator back to `src/main.rs` as the
`std::once::Once` in `tests/api.rs` will populate it anyway and it
doesn't make sense for it to be here.

* Add some basic tests

Add some basic tests under the `tests/` directory. There are only two
tests, which setup the server and ping two endpoints, one of which
requires a connection to the MongoDB server itself.

* Add unit tests for /api/users

Add more unit tests for /api/users routes. This includes failure
conditions, such as registering with the same email as a previous user
and logging in with the incorrect password.

This should be used in conjunction with a local database instance for
testing purposes. It will delete all stored data upon running `cargo
test`, before adding some test data. This can be done on the Atlas
instance, but may lead to some issues if people run tests at the same
time.

In the future, it may be easier to mock the MongoDB instance to mitigate
the need for this, however for now it should be fine to setup the
database and run some API queries on the data.

* Move dotenv::dotenv() back to main

Move the environment populator back to `src/main.rs` as the
`std::once::Once` in `tests/api.rs` will populate it anyway and it
doesn't make sense for it to be here.

* Return id.to_string() instead of id

Return the stringified version of id as it causes less nesting in the
resulting JSON.