Add documentation to everything
Add documentation to every publically accessible method that
`#![warn(missing_docs)]` complains about. Documentation can be viewed by
running `cargo doc --open` and should allow for easier development in
the future.

Add doctests, which can be run individually with `cargo test --doc` and
allow examples to be shown in the browser.

Enable some additional compiler warnings for missing documentation,
debug implementations and Rust 2018 idioms (although this is unlikely to
be needed).

Remove some unused methods that were previously used for testing.

In the future, documentation can be added to API routes to show the
expected input and output formats.