Add pbkdf2_iterations to the config
Add the `pbkdf2_iterations` variable to the configuration file, allowing
people to use different iteration counts depending on the environment.
This allows less iterations for tests for example, so they can be
completed quicker.

Remove the checking of hashes when adding a new user, as this is really
just a sanity check and wasn't doing anything with the result.