Add and refactor test suite
Add tests for the project related routes, verifying that endpoints deal
with requests in the way we expect them.

Refactor the API testing suite into 3 files, one dealing with the core
of the testing system such as database setup, and the other 2 dealing
with users and projects respectively.

Add constants for the ObjectIds used for testing, as this makes them
seem less like magic numbers and allows the tests to be a little more
descriptive.

Remove the definitions of some structs as they were copied straight
from the server itself, instead choosing to import them.