Infer dataset column types on addition
When the user adds a dataset to their project, infer the types of each
column into either Numerical or Categorical. Store this in the database
so that in future it can be shown on the dashboard for users.

Dataset column inference works by attempting to parse each value as a
float, if this does not work then the column is categorised as
Categorical data. By default we assume all columns are numerical.