Switch to using a Mutex instead of channels
Switch to using a queue with a Mutex around it instead of a channel
based approach due less chance of the structure becoming locked. The
previous approach could deadlock if no more data was being sent, meaning
the outbound thread would block trying to read from the channel, thus
preventing it from sending any data onwards.

The new approach has both threads continually reading and writing. While
this probably has bad lock contention, it should work for small
examples and getting things working.

If the DCL dies at any point and cannot receive messages, the queue will
begin to fill up, as the outbound thread will not be able to connect to
it. This is fine, as it means that we have some persistant storage if
the DCL goes down, identifiers will be sent across to it at the earliest
possible time.