Only allow one dataset per project
Previously, sending another dataset to this route would simply insert
another dataset into the database, instead of overwritting the existing
data. While the backend would assume that only one dataset existed for a
project, this could lead to unintended consequences in the future.

Thus, this checks whether the project already has data, and then deletes
it if it does before reinserting the sent values. This could be done
with upsert, however this would not return the inserted identifier,
changing the functionality of the route (and also making it more
complex).

Fixes #74.