Upgrade error handling
Upgrade the error handling for the `api-server`, making use of
`tide::Error` to return earlier from functions correctly. This also
changes some of the status codes to be more representative, such as
changing to a 401 Unauthorized for an invalid password.

Add helper functions for checking whether a user or project exists,
reducing code repetition and allowing the `?` operator to be used.