Infer train and predict data as one file upload (#113)
Infer the training and prediction data from the file upload given by the
user. This assumes the last column is the column that needs predicting
and splits the dataset based on whether this column is filled.

Datasets then store the training and prediction data as compressed in
the database, with both being returned when requested (but only the
training data shown currently).

This allows the machine learning to have some values to predict and thus
give results for.