Add beginning of DCL-DCN communication
Allow nodes to connect to the DCL to register themselves as a new model,
given a name and email address. They can then sign a challenge and
respond to the server with it. After this, they are expected to send
back the token that the server gave them, and the DCL will authenticate
them.

This corresponds to the model registration diagram drawn by Graham and
the initial stage of DCL communication drawn by Freddie. Currently, if
any communication is out-of-order, the thread will panic.