Remove `read_stream` function
Remove the `read_stream` function in favour of sized messages in the
distributed control layer. Instead of being null-terminated, messages
are prefixed with their size, allowing implementations to define the
length of a message and read a given number of bytes.

This also allows for streaming in future, where portions of a message
can be read and forwarded on which waiting for the next part.