Key nodes on model_id instead of access token
Key nodes in the DCL pool using their model identifier instead of their
access token, as this allows them to be uniquely identified alongside
telling us which models are running and when.

This can be used in the future to more easily track the work a model
does and its prediction results.