Add client-DCL communication protocols (#136)
* Add beginning of DCL-DCN communication

Allow nodes to connect to the DCL to register themselves as a new model,
given a name and email address. They can then sign a challenge and
respond to the server with it. After this, they are expected to send
back the token that the server gave them, and the DCL will authenticate
them.

This corresponds to the model registration diagram drawn by Graham and
the initial stage of DCL communication drawn by Freddie. Currently, if
any communication is out-of-order, the thread will panic.

* Update API server return messages

Update the messages the API server returns to always have a variant,
allowing them to be parsed more easily in the Python code.

* Use a more future-proof automaton style handler

When executing the protocol for the node end, use a more automaton-like
handler for messages. This allows it to read and store a single message,
before responding and waiting for the next one.

This allows it to deal with both the `AccessToken` first case and the
`NewModel` case.

* Refactor health checking

* Add basic mocking for tests with mockito

Use the `mockito` library to mock the responses from the API server.
This allows the DCL tests to run independently without the server
itself.

* Update health checking and job sending

Update the health checking to send the current timestamp alongside the
message itself, to ensure it has a variant. Add a config string to the
`JobConfig` variant to allow it to be represented more easily in Python
as well.

* Reduce CPU usage using a Condvar

Reduce the CPU usage of the `interface` while idle by using a `Condvar`,
which acts as a semaphore. This prevents it from spinlocking.

Fixes #131.

* Allow uploading of predictions to the database

Allow the DCL to upload prediction results to the database, under a new
collection called `predictions`. This is likely to be temporary, and
will be changed to a better schema in the future. Predictions are stored
with the project identifier, so they can be collated in future.

* Compress predictions before inserting them

Compress the predicted data before inserting into the database for more
consistency with other formats, and to save storage space.

* Add route to get predictions for a project

Add a route to get the predictions generated so far for a given project
identifier. This allows them to be displayed on the frontend for users
to see.

* Force predictions to have a Binary

Change the `predictions` field of `Prediction` to not be optional, as
predictions should always have something here. This then reduces the
requirement of unwraps elsewhere, and better represents the constraints.

* Add length prefixing to TCP messages

This fixes problems that are likely to occur in the future with larger
datasets or laggy clients. Messages are now expected to be prefixed with
4 big-endian bytes referring to the size of the message after it.

* Re-add timestamps to logging

* Remove `read_stream` function

Remove the `read_stream` function in favour of sized messages in the
distributed control layer. Instead of being null-terminated, messages
are prefixed with their size, allowing implementations to define the
length of a message and read a given number of bytes.

This also allows for streaming in future, where portions of a message
can be read and forwarded on which waiting for the next part.

* Improve logging and fix some tests

* Fix test builds

* Model Output View (#132)

* New View to Observer Model Results

* Updated header styling

* Styling and tab changes

* Project start to begin processing (#133)

* New View to Observer Model Results

* Updated header styling

* Styling and tab changes

* Begin processing for user

* Send a training and prediction dataset to nodes

Previously, we would only send the training dataset to the nodes for
processing, however we actually need to do some learning. We now send
both the training and the prediction dataset to nodes for jobs.

* Remove duplicate message type

* Add a better format for prediction responses

Prediction responses were previously just raw bytes, but this does not
work with the new format of messages. Instead, this adds a Message
variant for the predictions so they can be parsed correctly.

* Key nodes on model_id instead of access token

Key nodes in the DCL pool using their model identifier instead of their
access token, as this allows them to be uniquely identified alongside
telling us which models are running and when.

This can be used in the future to more easily track the work a model
does and its prediction results.

* Fix tests and remove guaranteed failures

Fix the unit tests to have a prediction dataset alongside the training
dataset and remove some tests that require mocking for completion (or
the `api-server`) to be running.

* Remove mockito and update dependencies

* Add the mallus subdirectory to dodona

* Updating mallus

* Fully remove `mockito`

Fully remove `mockito` as a dependency as the mocking does not work for
integration tests.

* Remove a conditional compile attribute

Remove a `#[cfg(not(test))]` attribute as this was used for mocking
which doesn't work.

* Improve the pylint score

* Style updates (#138)

* New projects update when submitted

* Project updates when started

* Stylistic Updates

Co-authored-by: Freddie Brown <34427683+FreddieBrown@users.noreply.github.com>

Co-authored-by: Nathan Hall <37386274+hallnath1@users.noreply.github.com>
Co-authored-by: Graham Kemp <grahamk206@gmail.com>
Co-authored-by: Freddie Brown <34427683+FreddieBrown@users.noreply.github.com>