Avoid allocations for status enums (#154)
When updating the status of a project or model, this required us to have
a custom impl for `std::fmt::Display`, which would allocate a new
string each time. Instead, we can impl `From<State>` for `bson::Bson`
and use strings in the binaries.

In reality, this makes no difference to performance, but is a little
more ergonomic to use when changing the states.