Add tests for the DCL-client protocol
Implement and validate some basic tests for the DCL to client
communication protocol. Tests use `mockito` for simulating and mocking
the responses of the API server, which works as these run as unit tests
instead of integration tests.

This checks that clients can immediately send their access token,
invalid messages and incorrect orderings fail and that the full
end-to-end protocol of creating a model, sending a challenge response
and returning an access token will work.