Add tests for DCL-client protocol (#162)
* Optimise the tests to be faster

Reduce the run time of the tests by sleeping for 1/10th of the time.
Tests still work identically, just take less time to run compared to
before.

* Add tests for the DCL-client protocol

Implement and validate some basic tests for the DCL to client
communication protocol. Tests use `mockito` for simulating and mocking
the responses of the API server, which works as these run as unit tests
instead of integration tests.

This checks that clients can immediately send their access token,
invalid messages and incorrect orderings fail and that the full
end-to-end protocol of creating a model, sending a challenge response
and returning an access token will work.

* Move the node protocol to a subdirectory

Move the node protocol handling to its own submodule in `node_end` to
allow its tests to be better named.