Change to using blocking async_std for MongoDB
As the API server and DCL both rely on the `async_std` feature for
`mongodb`, using the `sync` feature in the interface will cause the
tests to fail due to cross compilation. Instead, the easy fix for the
moment is to use the `async_std` feature but block on all asynchronous
methods.

In the future, the interface should be made asynchronous (and hopefully
on Tokio across the board) and this should be revisited.