Fix #134 (#169)
* Extract interface forwarding to separate function

Extract the forwarding of dataset identifiers to the interface into a
new function to allow it to be more self-contained.

* Insert identifiers upon forward failures

When forwarding to the interface, check whether the operation fails or
not. If it does, attempt to insert the identifier into a MongoDB
collection for persistent storage.

* Add a struct for interface jobs

Add a struct to manage the structure of interface jobs.

* Rename collection name to be simpler

* Allow the interface to read from Mongo on startup

Allow the interface layer to read from the MongoDB instance job queue
when it begins execution to ensure that job requests are not missed (at
least not as often).

* Change to using blocking async_std for MongoDB

As the API server and DCL both rely on the `async_std` feature for
`mongodb`, using the `sync` feature in the interface will cause the
tests to fail due to cross compilation. Instead, the easy fix for the
moment is to use the `async_std` feature but block on all asynchronous
methods.

In the future, the interface should be made asynchronous (and hopefully
on Tokio across the board) and this should be revisited.