Refactor the length prefixing into a trait
Length prefixing for messages previously shared an implementation
between both message types. This left it vulnerable to one being changed
without the update being propagated.

Instead, this adds a `ReadLengthPrefix` trait for reading a length
prefixed message from a stream and a `WriteLengthPrefix` trait for
converting a type into a vector of bytes prefixed with its length.

In some cases, we want to length prefix a message but not contain the
type information, such as when acting as a JSON proxy for the API
server. For this, the `RawMessage` type acts as a wrapper, only
serializing its contents.