Job config (#175)
* Extract interface forwarding to separate function

Extract the forwarding of dataset identifiers to the interface into a
new function to allow it to be more self-contained.

* Insert identifiers upon forward failures

When forwarding to the interface, check whether the operation fails or
not. If it does, attempt to insert the identifier into a MongoDB
collection for persistent storage.

* Add a struct for interface jobs

Add a struct to manage the structure of interface jobs.

* Rename collection name to be simpler

* Eof check (#168)

* split protocol into own mod

* created auth_end

* fixed EOF issue

* created ClientMessage and InterfaceMessage

* Extract interface forwarding to separate function

Extract the forwarding of dataset identifiers to the interface into a
new function to allow it to be more self-contained.

* Insert identifiers upon forward failures

When forwarding to the interface, check whether the operation fails or
not. If it does, attempt to insert the identifier into a MongoDB
collection for persistent storage.

* Add a struct for interface jobs

Add a struct to manage the structure of interface jobs.

* created ClientMessage and InterfaceMessage

* Integrated InterfaceMessage

changing so that all messages over interface should be interface
messages. Currently only a Config message but has capacity for
more in the future.

* writing config to node

* Add comments

* added inferred columns to configs

* sending proper columns types to client

* written frontend job config dropdown

* Upgrade the DCL to use Tokio 1.0

* Use a GitHub version of MongoDB

* Update all Tokio versions to 1.0

* Remove DCL specific messages

Remove the DCL specific messages as the messages module is updated to
Tokio 1.0 now.

* Update interface to use Tokio 1.0

Update the interface to be asynchronous, using Tokio 1.0 for its runtime
inline with most other components.

* Change the interface to use channels

Change the interface layer to use Tokio's channels instead of a queue,
as the blocking Mutex causes task execution to stop.

* Prefer tokio::time::sleep over thread::sleep

Asynchronous processes will be blocked if using thread::sleep as they
run on a singular thread. Instead, tokio::time::sleep does the same
thing asynchronously and thus allows other tasks to make progress.

* Clean up logging output slightly

* Api tokio into job_config (#178)

* converted most to actix

* fixed common test stuff

* tests don't work and it panics

* added actix and changed rt version

* added arcs

* solved filter issue

* error implementing factory thing

* add await

* converted code base to use other route reg

* getting tests working

* projects tests all pass

* tests all pass

* finalising api_tokio changes

* Upgrade Tokio 0.3 to 1.0

* Remove tokio-util as it isn't needed

* Upgrade all dependencies to be Tokio 1.0

Upgrade all the `actix-*` dependencies to be compatible with Tokio 1.0
by using beta versions, and use the Git dependency for MongoDB for now.

* Change DodonaError::name to return &'static str

Change the `DodonaError::name` to return a `&'static str` instead of a
String as allocations are not needed.

* Refactor the length prefixing into a trait (#180)

Length prefixing for messages previously shared an implementation
between both message types. This left it vulnerable to one being changed
without the update being propagated.

Instead, this adds a `ReadLengthPrefix` trait for reading a length
prefixed message from a stream and a `WriteLengthPrefix` trait for
converting a type into a vector of bytes prefixed with its length.

In some cases, we want to length prefix a message but not contain the
type information, such as when acting as a JSON proxy for the API
server. For this, the `RawMessage` type acts as a wrapper, only
serializing its contents.

* made building a cluster more resiliant and mdb tokio 1.0 (#181)

* made building a cluster more resiliant

* cleaning up

* Fix a lot of `clippy` warnings

* Address review comments

* Fix merge issues

Co-authored-by: Alexander Jackson <alexanderjackson@protonmail.com>