Allow basic constraints to be specified
Now that the API server can ensure we know who a user is, it's worth
checking whether the given user has permissions for that project for
example. This can be done through the `check_user_owns_project`
function, which returns a 422 on invalid ObjectId, 404 on failing to
find the project, 422 on failing to deserialize the MongoDB data and 401
if the user does not own the project.