Refactor models to be more type-safe
Refactor the MongoDB models structure to better constrain the data we
expect to be storing. For example, all `_id` fields are now `ObjectId`
instead of `Option<ObjectId`, as all instances should have an
identifier. These are created in the constructor, so there's no
interaction with MongoDB.

This also removes a lot of `unwrap` and `expect` calls, where these were
needed for conditions such as "project doesn't have a user". Instead,
projects must be created with a user.