Change all `{user_id}` routes to be authorised
Change all the routes containing `{user_id}` as a parameter to instead
take an `auth::User` as a parameter to their handler. This means the
parameter is no longer needed, so the frontend query routes must be
updated as well. This also removes the need for some checks, as users
must exist to be authorised at the route level.

This also breaks the frontend routes almost completely, so this will
need to be changed when the API server returns a JWT on successful
registration and logins.