Return JWT tokens on login
When the user logs in successfully, create them a JWT token using the
server's private key and return it to them. This is then stored in the
cookies, along with being sent on every Axios request. When the user
signs out, remove the token from their cookies and delete it from the
Axios headers.

Fixes #47.