Add authentication to the API server and frontend (#208)
* Add basic JWT authorisation

Implement a basic JWT authorisation scheme. For the moment, only one
route requires authorisation and users have no way of getting a signed
token. This is done through the `auth::User` type, which can create
itself from the incoming HTTP request or return an Unauthorized error
message.

This works similarly to the concept of Rocket's guards, where functions
that require additional checks can have an `auth::User` as an input
parameter.

Tests have been added to ensure the implementation functions correctly,
failing on conditions such as missing headers, invalid headers, expired
tokens and incorrect keys used for signing. There is also a test to
ensure that correct JWTs should be accepted.

* Allow configuration of the JWT key and fix tests

Allow the JWT key used for authentication to be configurable by using
the `backend/api-server/jwt_key` file which gets included on
compilation. This also allows the tests to more easily access it and
fixes the tests that were previously failing due to lack of
authentication.

* Allow basic constraints to be specified

Now that the API server can ensure we know who a user is, it's worth
checking whether the given user has permissions for that project for
example. This can be done through the `check_user_owns_project`
function, which returns a 422 on invalid ObjectId, 404 on failing to
find the project, 422 on failing to deserialize the MongoDB data and 401
if the user does not own the project.

* Change all `{user_id}` routes to be authorised

Change all the routes containing `{user_id}` as a parameter to instead
take an `auth::User` as a parameter to their handler. This means the
parameter is no longer needed, so the frontend query routes must be
updated as well. This also removes the need for some checks, as users
must exist to be authorised at the route level.

This also breaks the frontend routes almost completely, so this will
need to be changed when the API server returns a JWT on successful
registration and logins.

* Return JWT tokens on login

When the user logs in successfully, create them a JWT token using the
server's private key and return it to them. This is then stored in the
cookies, along with being sent on every Axios request. When the user
signs out, remove the token from their cookies and delete it from the
Axios headers.

Fixes #47.

* Add authorisation to all remaining routes

Implement authorisation for all the remaining frontend facing routes and
update the tests accordingly. DCL facing routes remain unchanged and do
not need a valid JWT to access for the time being. Check some
preconditions on users owning projects while we are at it.

* Rename `auth::User` to `auth::Claims`

Rename the `auth::User` struct to `auth::Claims` for 2 reasons. Firstly,
it better represents what the type contains, which is the claims that a
request makes about its sender. It also clashes less with the
`models::User` type and thus is a bit easier to read.

* Add basic JWT authorisation

Implement a basic JWT authorisation scheme. For the moment, only one
route requires authorisation and users have no way of getting a signed
token. This is done through the `auth::User` type, which can create
itself from the incoming HTTP request or return an Unauthorized error
message.

This works similarly to the concept of Rocket's guards, where functions
that require additional checks can have an `auth::User` as an input
parameter.

Tests have been added to ensure the implementation functions correctly,
failing on conditions such as missing headers, invalid headers, expired
tokens and incorrect keys used for signing. There is also a test to
ensure that correct JWTs should be accepted.

* Allow configuration of the JWT key and fix tests

Allow the JWT key used for authentication to be configurable by using
the `backend/api-server/jwt_key` file which gets included on
compilation. This also allows the tests to more easily access it and
fixes the tests that were previously failing due to lack of
authentication.

* Allow basic constraints to be specified

Now that the API server can ensure we know who a user is, it's worth
checking whether the given user has permissions for that project for
example. This can be done through the `check_user_owns_project`
function, which returns a 422 on invalid ObjectId, 404 on failing to
find the project, 422 on failing to deserialize the MongoDB data and 401
if the user does not own the project.

* Change all `{user_id}` routes to be authorised

Change all the routes containing `{user_id}` as a parameter to instead
take an `auth::User` as a parameter to their handler. This means the
parameter is no longer needed, so the frontend query routes must be
updated as well. This also removes the need for some checks, as users
must exist to be authorised at the route level.

This also breaks the frontend routes almost completely, so this will
need to be changed when the API server returns a JWT on successful
registration and logins.

* Return JWT tokens on login

When the user logs in successfully, create them a JWT token using the
server's private key and return it to them. This is then stored in the
cookies, along with being sent on every Axios request. When the user
signs out, remove the token from their cookies and delete it from the
Axios headers.

Fixes #47.

* Add authorisation to all remaining routes

Implement authorisation for all the remaining frontend facing routes and
update the tests accordingly. DCL facing routes remain unchanged and do
not need a valid JWT to access for the time being. Check some
preconditions on users owning projects while we are at it.

* Rename `auth::User` to `auth::Claims`

Rename the `auth::User` struct to `auth::Claims` for 2 reasons. Firstly,
it better represents what the type contains, which is the claims that a
request makes about its sender. It also clashes less with the
`models::User` type and thus is a bit easier to read.

* Make finishing touches and fix merge issues

Make some finishing touches in terms of naming conventions and
documentation, alongside fixing some issues that came from the merge
with `model-refactor`.

* Return JWTs on registration

When the user registers successfully, perform the same process as we do
on login by creating a JWT for them and setting it as an Axios
parameter.

* Use a global Axios instance for sending headers

Previously, the frontend would send authorised requests until you
refreshed the page, at which point it wouldn't send an Authorization
header at all. Instead, this will continue to send them across page
reloads without too many changes.

* Replace all `axios.` with `this.$http.`

Use the global `axios` instance for all requests as per Nathan's
comments.