Simplify the tests using a macro to build the API (#214)
A lot of the code in the tests is repeated, getting the state for the
API and then setting up the test server to send requests at. This is
also quite hard to read, and discourages people from adding more tests.

Instead, we can use a simple macro to build the API and fetch the state
each time. This takes the method, route and handler and builds the API
for you, removing ~10 lines of code each time it's used and making it
much easier to see which routes are being used.

Hopefully a similar approach is possible for sending requests, at which
point testing the API can be a lot better structured.