Allow JWTs to expire (#220)
Previously, JWTs would expire technically speaking, but only in the year
292,277,026,596 (292 billion). By this point, lots of interesting things
would have happened [1], such as the evolution of the Sun into a red
giant, Jupiter potentially causing Mercury to collide with Venus and
the concept of time becoming a little distorted. In 50,000 years we
will need a leap second every day just to keep up, so imagine what it
will be like when these tokens expire. We'd also have collided with the
Andromeda galaxy (and numerous other issues), but at least the tokens
are still good to go.

Anyway, tokens expire after 2 weeks now. Fixes #210.

[1]: https://en.wikipedia.org/wiki/Timeline_of_the_far_future