Update the MongoDB dependencies to be official (#221)
* Update the MongoDB dependencies to be official

Update our dependencies on MongoDB to use the `master` branch of the
official repository, which recently upgraded to Tokio 1. An official
release seems to be version 2, which could potentially take a while to
come out.

* Use the MongoDB v2 alpha release instead

Mongo actually released a v2.0.0-alpha version to `crates.io`, so just
use that instead for now.