Use the `float-cmp` crate to compare floats
The GitHub CI seems to break, as it compares 0.999999 to 1 and finds
they are not equal. Instead of using strict equality, `float-cmp` adds
some lenience into the comparison. While the tests pass locally for me
using either method, this is technically more correct and should work in
the CI builds as well.