Add Job Config Form
Add the problem type - classification or regressions to job config
form and the prediction column so does not have to be inferred.
Remove the dropdown for job config making it permanent
Moved the page around a bit moving the job config form up
Turned the link into a button and removed the background so
disabling it works better.