Begin adding large file support
Currently, users are constrained to approximately 32Kb of data in their
uploaded datasets. Realistically, many users are going to want to
process huge amounts more, so we need to allow larger upload sizes.

The easiest way to do this is through the multipart form specification.
This allows unlimited upload sizes, as long as the web server can
continue to handle it. However, we then need to actually store the files
in a viable way for future processing.

MongoDB advises against having large documents in the database, with the
theoretical boundary being 16MB. While MongoDB has large file support in
the form of GridFS, the current driver does not support it. Instead, we
can make our own GridFS-like models, which contain a `File` and a
`Chunk` for streaming data. This allows us to store large files
efficiently and also process them through streaming later.

Currently, this is very experimental. The API server will happily upload
the data, but no additional processing is done and they are not linked
well to the DCL and other components. Instead, a blank dataset will be
stored where other components expect to find it.