Fix the tests and impl message traits generically
Fix the `interface` test in the DCL that was breaking as it was sending
a `JobConfiguration` but the DCL was expecting a `Job` instead. The
`Job` currently only contains the identifier alongside the
configuration, but it may be useful in future to have that alongside it.

This was made easier by implementing the `ReadLengthPrefix` and
`WriteLengthPrefix` traits generically, over anything that can be
deserialized and serialized. In theory, this allows basically any type
to be sent across a network now and is a lot more flexible.