Update Job config (#257)
Job config now contains A prediction Column and a prediction type - Classification or regression

* Remove unused_macros and dead_code warnings

Added labels to ignore the warnings produced by api_with and
NON_EXISTENT_PROJECT_ID as these are both used elsewhere but
the compiler does not pick them up. Added allow tags for both
of these to remove the warnings

* Add Job Config Form

Add the problem type - classification or regressions to job config
form and the prediction column so does not have to be inferred.
Remove the dropdown for job config making it permanent
Moved the page around a bit moving the job config form up
Turned the link into a button and removed the background so
disabling it works better.

* Switches to Yarn from NPM

Probably shouldn't be in this branch but it got annoying
Changes the docker system to use yarn by removing package-lock
replacing it with yarn.lock. Also adds chart.js to package.json

* Updates config.toml to match the default

* Replace package-lock with yarn.lock

* Add prediction_column and prediction_type to JobConfig

Update the interface message with a Prediction Column as String
Add a PredictionType enum which is currently Classification or
Regression.
Updates all instaces of InteraceMessage Config and Client Message
with the new types

* Rename variable to convention

* Allow processing to take `u32` instead of `str` (#256)

When uploading the job configuration, allow a raw JSON integer instead
of forcing it to be a string by using a `struct` instead.

Add a test to ensure this will continue to work.

Fixes #196.

* Update mod.rs

* fix dcl tests

* Add interface listen to config.toml

Co-authored-by: Alexander Jackson <alexanderjackson@protonmail.com>