Make jobs persistent in the database (#252)
* Make jobs persistent in the database

Previously when sending a job from the `api-server` to the `interface`,
the configuration settings would only be stored within the database if
the message failed to send. This was so that if the `interface` was
down, it would be able to pick up the jobs it missed when it restarted,
reducing the chance of jobs going missing.

However, if we want to allow users to reuse a previous configuration,
this meant we had to hope the `interface` was down when sending,
otherwise we would have no record of it. This also reduced the logging
of the application slightly, as we wouldn't be able to track what had
happened.

This ensures that all jobs are stored within the database before being
sent to the `interface` layer. When they have been sent on from the
`interface`, they will be marked as `processed` in the database so they
are not re-sent at a later date.

The `InterfaceMessage` enum has been removed, as it served only to send
the configuration and had no other variants. Instead, the `api-server`
will create an instance of a `JobConfiguration`, which will be stored
with the related `Job` in the database. The DCL still only receives the
`JobConfiguration` as before, but the `interface` requires the
identifier of the `Job` to mark it as processed.

* Fix the tests and impl message traits generically

Fix the `interface` test in the DCL that was breaking as it was sending
a `JobConfiguration` but the DCL was expecting a `Job` instead. The
`Job` currently only contains the identifier alongside the
configuration, but it may be useful in future to have that alongside it.

This was made easier by implementing the `ReadLengthPrefix` and
`WriteLengthPrefix` traits generically, over anything that can be
deserialized and serialized. In theory, this allows basically any type
to be sent across a network now and is a lot more flexible.

* Add a timestamp to `Job`s

When creating a new job, store the time it was created at. In future,
this will allow us to query the database for the last job that a user
ran on a given dataset, meaning we can load the configuration in for
them.