Client perf (#261)
* written model performance endpoint

* can get data, can't render

* building card

* showing performance on frontend

* chart renders without the available data

* added BarChart

* Fix chart on collapse (#260)

* Format and when collapse is shown call child component event

Add references to the model-performance component which contains
the i value of the ModelCard. Added a event handler for @shown
which when the collapse is activated calls the show method on the
model-performance component

* Will load the graph is not previously loaded

Adds a chartLoaded bool to data which is initialized as false
When this component receives the event from the parent component
checks if chart has been loaded, if not calls renderChartOnCollapse

* Move renderChart to function

Rather than rendering the cahrt on mounted the chart is rendered
through a function. This function is called when the parent
collapse is shown.

* can dynamically render performance data

* nice charts and removed annoying otp text

* added test and sort by newest

* removing warning

* Update web/src/components/ModelCard.vue

Co-authored-by: Graham Kemp <44637230+G-Kemp101@users.noreply.github.com>

* tests should be more resilliant

* Iterate once when getting performances

When getting the performances of each model, just iterate once instead
of collecting and then iterating again.

Co-authored-by: Graham Kemp <44637230+G-Kemp101@users.noreply.github.com>
Co-authored-by: Alexander Jackson <alexanderjackson@protonmail.com>