Swap to `fern` for logging and minimise output
Follow in the footsteps of the other components and use `fern` for
logging. As the `api-server` logs `actix-*` components as well as
itself, add a method to create a logger with different levels for
modules other than the crate itself.

Minimise the output of the logging to be a little cleaner. For the
moment, we don't care where requests come from and we aren't interested
in their User Agents, so just log the response code and the first line
of the request, which includes the method and route.