Refactor the error type to `ServerError`
Rename the error type to `ServerError` and move it to `src/error.rs`.
Add some type aliases for it, to allow for using `ServerResult<T>` when
returning a given type and also handling errors and `ServerResponse` for
the general `Result<HttpResponse, ServerError>`.

This shortens some of the routes a little as `rustfmt` doesn't need to
wrap lines as often.